requirements/requirements-testing.txt
# PyTest for running the tests.
pytest==3.1.3
pytest-django==3.1.2
pytest-cov==2.5.1
pytest-selenium==1.11.0
psycopg2
factory-boy==2.8.1
git+https://github.com/heoga/django-mock-queries.git@django11